Remove procurement.order
